working omsu schedule

+++ src/main/java/dev/mvr/schedule/repository/StudentRepository.java
@@ -8,15 +8,15 @@ import java.util.List;
 import java.util.Map;
 
 public class StudentRepository {
-    private static Map<Integer, List<UniversityGroup>> studentGroup = new HashMap<>();
+    private static Map<Long, List<UniversityGroup>> studentGroup = new HashMap<>();
 
-    public static List<UniversityGroup> getStudentGroups(int id){
+    public static List<UniversityGroup> getStudentGroups(long id){
         return studentGroup.getOrDefault(id,new ArrayList<>());
     }
-    public static void addGroup(int studentId, UniversityGroup universityGroup){
+    public static void addGroup(long studentId, UniversityGroup universityGroup){
         studentGroup.getOrDefault(studentId,new ArrayList<>()).add(universityGroup);
     }
-    public static void addStudent(int studentId){
+    public static void addStudent(long studentId){
         studentGroup.put(studentId,new ArrayList<>());
     }
 }
